1.56 реорганизую диалоги
