+++ app/src/main/java/com/pavlov/nearWarSecrets/theme/uiComponents/StyledDialog.kt
@@ -1,0 +1,38 @@
+package com.pavlov.nearWarSecrets.theme.uiComponents
+
+import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Surface
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.window.Dialog
+
+@Composable
+fun MyStyledDialog(
+    onDismissRequest: () -> Unit,
+    content: @Composable () -> Unit
+) {
+    Dialog(onDismissRequest = onDismissRequest) {
+        Surface(
+            shape = RoundedCornerShape(8.dp),
+            color = MaterialTheme.colors.surface,
+            modifier = Modifier
+                .border(
+                    width = 2.dp,
+                    color = Color.Green,
+                    shape = RoundedCornerShape(8.dp)
+                )
+        ) {
+            Column(
+                modifier = Modifier.padding(16.dp)
+            ) {
+                content()
+            }
+        }
+    }
+}
