+++ app/src/main/java/com/pavlov/nearWarSecrets/ui/Images/extracted/saveImageToPrivateStorage.kt
@@ -1,4 +1,4 @@
-package com.pavlov.nearWarSecrets.ui.Images
+package com.pavlov.nearWarSecrets.ui.Images.extracted
 
 import android.content.Context
 import android.net.Uri
